+++ run_tests.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
-cd /Users/carltanner/dev/ruby/gems/placekey_rails
+cd /users/carltanner/dev/ruby/gems/placekey_rails
 bundle exec rspec